ml-service/requirements.txt
fastapi
uvicorn
redis
kafka-python
pydantic
pydantic-settings
python-dotenv